experiments/exp04_surveyforge/requirements.txt
# experiments/exp04_surveyforge/requirements.txt
# Зависимости SurveyForge (из repos/SurveyForge/code/requirement.txt + наши утилиты)
#
# CUDA: образ pytorch/pytorch:2.2.0-cuda12.1-cudnn8-runtime уже содержит torch с CUDA.
# faiss-gpu: совместима с CUDA 11.x/12.x; если GPU недоступен — замени на faiss-cpu.

# --- torch (уже в base image, но фиксируем версию) ---
# torch==2.2.0  # раскомментируй если используешь другой base image

# --- Embeddings & retrieval ---
transformers==4.44.2
sentence-transformers==2.7.0
accelerate==0.33.0
safetensors==0.4.4
faiss-cpu==1.9.0        # замени на faiss-gpu==1.7.2 при наличии CUDA

# --- LangChain (нужен SurveyForge) ---
langchain==0.3.4
langchain-community==0.3.0
langchain-core==0.3.12
langchain-huggingface==0.1.0
langchain-openai==0.2.3
langchain-text-splitters==0.3.0

# --- LLM API ---
openai>=1.0
anthropic>=0.20

# --- Data & storage ---
tinydb==4.8.0
pandas>=2.0
numpy>=1.23
scipy>=1.10
scikit-learn>=1.3
tiktoken>=0.5
tokenizers>=0.19
nltk>=3.9

# --- Utils ---
PyYAML>=6.0
python-dotenv>=1.0
tqdm>=4.66
requests>=2.31
httpx>=0.27
aiohttp>=3.10
pydantic>=2.5
huggingface_hub>=0.20
